♻️ 세미나 페이지 getpost 리팩토링
